[MATTER-446][K32W0] Add delay when rebooting after OTA finishes
A delay is needed in order to ensure that all pending actions
are actually executed before rebooting with the new image. For
example, if someone is subscribed to an event, some events may
be missed by the subscriber: state transition from Downloading
to Applying when doing OTA. The delay can be overwritten by
setting the define to a new value.

Signed-off-by: Marius Tache <marius.tache@nxp.com>